Rework blog post selection code for localization
Previously, we were able to select the “Drafts” tab by name without localizing it. It’s now necessary to choose it by its localized string, so now that works.

In some cases, it appears that fetching the correct localization may not be reliable, so we only select it on the first run, and choose the post by name without reloading the posts list each time.

This necessitated adding a `scrollToTop` method to the `PostListPage` – it allows us to reset the list each time so that we can scroll down to the correct post, even if it’s above the last one we took a screenshot of.